Redesign program structure
