+++ .gitignore
@@ -1,7 +1,6 @@
 .stack-work/
 *~
-lowarn.cabal
-lowarn-programs.cabal
+*.cabal
 
 *.o
 *.hi
