+++ lowarn/app/Main.hs
@@ -1,23 +1,20 @@
 module Main (main) where
 
-import qualified Data.Sequence as Seq
-import DynamicLinker (load)
-import Types (Program (..))
+import Lowarn.DynamicLinker (load)
+import Lowarn.Types (Program (..))
 
-loadProgram :: String -> IO (a -> IO b)
-loadProgram moduleName = do
+loadProgram :: String -> a -> IO b
+loadProgram moduleName oldState = do
   status <- load moduleName "program"
   case status of
-    Just (Program p) -> do
-      return p
+    Just (Program p t) -> do
+      let newState = t oldState
+      p newState
     Nothing ->
       error ("Loading " <> moduleName <> " failed")
 
 main :: IO ()
 main = do
-  program1 <- loadProgram "Program1"
-  users <- program1 ()
-  let users' = Seq.fromList users
-  program2 <- loadProgram "Program2"
-  () <- program2 users'
+  users <- loadProgram "Lowarn.Programs.Program1" ()
+  () <- loadProgram "Lowarn.Programs.Program2" users
   return ()
