+++ lowarn/src/Lowarn/Types.hs
@@ -1,0 +1,9 @@
+{-# LANGUAGE RankNTypes #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+
+module Lowarn.Types (Program (..)) where
+
+data Program a b c = Program
+  { program :: a -> IO b,
+    transformer :: c -> a
+  }
