+++ programs/program1/package.yaml
@@ -1,4 +1,4 @@
-name:                lowarn-programs
+name:                lowarn-programs-program1
 version:             0.1.0.0
 github:              "jonathanjameswatson/lowarn"
 license:             BSD3
@@ -6,20 +6,11 @@ author:              "Jonathan Watson"
 maintainer:          "23344719+jonathanjameswatson@users.noreply.github.com"
 copyright:           "2022 Jonathan Watson"
 
-extra-source-files:
-- README.md
-
-# Metadata used when publishing your package
-# synopsis:            Short description of your package
-# category:            Web
-
-# To avoid duplicated efforts in documentation and dealing with the
-# complications of embedding Haddock markup inside cabal files, it is
-# common to point users to the README.md file.
-description:         Please see the README on GitHub at <https://github.com/jonathanjameswatson/lowarn#readme>
+description:         Lowarn demo program 1
 
 dependencies:
 - base >= 4.7 && < 5
+- lowarn
 - containers
 
 ghc-options:
